fix: address pre-PR review findings (9 items from 5 agents)
