+++ cli/testdata/compose_digest_pins.yml
@@ -10,7 +10,7 @@ services:
     image: busybox:1.37-musl@sha256:19b646668802469d968a05342a601e78da4322a414a7c09b1c9ee25165042138
     volumes:
       - synthorg-data:/data
-    command: ["sh", "-c", "mkdir -p /data/logs /data/memory && chown -R 65532:65532 /data"]
+    command: ["sh", "-c", "set -e; mkdir -p /data/logs /data/memory; chown -R 65532:65532 /data"]
     cap_drop:
       - ALL
     cap_add:
