+++ docker/compose.yml
@@ -50,7 +50,7 @@ services:
     image: busybox:1.37-musl@sha256:19b646668802469d968a05342a601e78da4322a414a7c09b1c9ee25165042138
     volumes:
       - synthorg-data:/data
-    command: ["sh", "-c", "mkdir -p /data/logs /data/memory && chown -R 65532:65532 /data"]
+    command: ["sh", "-c", "set -e; mkdir -p /data/logs /data/memory; chown -R 65532:65532 /data"]
     cap_drop:
       - ALL
     cap_add:
@@ -133,7 +133,7 @@ services:
   nats:
     # renovate: datasource=docker depName=dhi.io/nats
     image: dhi.io/nats:2.12-debian13@sha256:89f40ed3ce7f69ad0af95c264368d3de41d72de604c97e7f1f8bd57d3a2b617f
-    command: ["-js", "-sd", "/data", "-m", "8222"]
+    command: ["-js", "-sd", "/data", "-m", "8222", "--max_payload", "4MB"]
     ports:
       - "${NATS_CLIENT_PORT:-3003}:4222"
     volumes:
